+++ plugins/medusa/medusa-custom-payments-react/src/index.ts
@@ -22,6 +22,8 @@ export { StripeWrapper } from "./connectors/stripe/StripeWrapper";
 export { PayPalWrapper } from "./connectors/paypal/PayPalWrapper";
 export { GlobalPayWrapper } from "./connectors/globalpay/GlobalPayWrapper";
 export { MollieWrapper } from "./connectors/mollie/MollieWrapper";
+export { MollieKlarnaForm } from "./connectors/mollie/MollieKlarnaForm";
+export type { MollieKlarnaBilling } from "./connectors/mollie/MollieKlarnaForm";
 export { BraintreeWrapper } from "./connectors/braintree/BraintreeWrapper";
 export type {
   BraintreeWalletType,